requirements.txt
opencv-contrib-python>=4.5.3
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.2
pillow>=8.2.0
tk>=0.1.0